requirements.txt
langchain
langchain-groq
langchain-google-genai
langchain-mcp-adapters
langchain-community
langchain-core
langchain_tavily
langchain-experimental
langchain-google-community[places]
langgraph
langgraph-checkpoint-sqlite
pydantic
requests
fastapi
python-dotenv
streamlit
uvicorn
PyYAML
httpx
-e .